Fixed an overflow in the mean length calculation
Former-commit-id: 4b67ddc0cc3387a4fded37e5f181ca3aa1d54c02

+++ public/java/src/uk/ac/ox/well/indiana/utils/sequence/SequenceUtils.java
@@ -363,15 +363,15 @@ public class SequenceUtils {
      * @return  the mean sequence length observed
      */
     public static float meanLength(Collection<? extends CharSequence> sequences) {
-        int length = 0;
-        int numSequences = 0;
+        float length = 0;
+        float numSequences = 0;
 
         for (CharSequence seq : sequences) {
             length += seq.length();
             numSequences++;
         }
 
-        return (float) length / (float) numSequences;
+        return length / numSequences;
     }
 
     private static final byte[] nucleotides = {'A', 'C', 'G', 'T'};
